+++ lib/go/api_queries.go
@@ -143,10 +143,6 @@ func (a *QueriesApiService) QueryStream(body QueryRequest) (string, *http.Respon
         log.Fatal("missing required argument ApiKey")
     }
 	localVarHeaderParams["authorization"] = "ApiKey " + localVarHttpHeaderApiKey
-    
-    // version
-    localVarVersion := a.Client.selectVersion()
-	localVarHeaderParams["x-rockset-version"] = localVarVersion 
 
 	// body params
 	localVarPostBody = &body