+++ lib/go/client.go
@@ -39,7 +39,6 @@ var (
 type APIClient struct {
 	cfg    *Configuration
     apiKey string
-    version string
 }
 
 type Service struct {
@@ -48,7 +47,7 @@ type Service struct {
 
 // Client creates a new API client. Requires a userAgent string describing your application.
 // optionally a custom http.Client to allow for advanced features such as caching.
-func ApiClient(cfg *Configuration, apiKey string, Version string) *APIClient {
+func ApiClient(cfg *Configuration, apiKey string) *APIClient {
 	if cfg.HTTPClient == nil {
 		cfg.HTTPClient = http.DefaultClient
 	}
@@ -56,7 +55,6 @@ func ApiClient(cfg *Configuration, apiKey string, Version string) *APIClient {
     c := &APIClient{}
     c.cfg = cfg
     c.apiKey = apiKey
-    c.version = Version
 	return c
 }
 
@@ -78,11 +76,6 @@ func selectHeaderContentType(contentTypes []string) string {
 // selectHeaderAccept join all accept types and return
 func (c *APIClient) selectHeaderAuthorization(authorization string) string {
 	return c.apiKey
-}
-
-// selectVersion return version 
-func (c *APIClient) selectVersion() string {
-	return c.version
 }
 
 // contains is a case insenstive match, finding needle in a haystack
@@ -250,6 +243,9 @@ func (c *APIClient) prepareRequest(
 
 	// Add the user agent to the request.
 	localVarRequest.Header.Add("User-Agent", c.cfg.UserAgent)
+    
+    // Add the version to the request header
+    localVarRequest.Header.Add("x-rockset-version", c.cfg.Version)
 
 	for header, value := range c.cfg.DefaultHeader {
 		localVarRequest.Header.Add(header, value)