+++ lib/go/model_source_file_upload.go
@@ -15,24 +15,15 @@ import (
     
 )
 
-// Integrations that can be associated with data sources to create collections. Only one type of integration may be specified.
-type Integration struct {
-	// ISO-8601 date
-	CreatedAt string `json:"created_at,omitempty"`
-	// email of user who created the integration
-	CreatedBy string `json:"created_by"`
-	// descriptive label and unique identifier
-	Name string `json:"name"`
-	// longer explanation for the integration
-	Description string `json:"description,omitempty"`
-	// credentials for an AWS key integration
-	Aws *AwsKeyIntegration `json:"aws,omitempty"`
-	// details of an AWS External Id integration
-	AwsExternalId *AwsExternalIdIntegration `json:"aws_external_id,omitempty"`
-	// details of a GCP Service Account integration
-	GcpServiceAccount *GcpServiceAccount `json:"gcp_service_account,omitempty"`
+type SourceFileUpload struct {
+	// name of the file
+	FileName string `json:"file_name"`
+	// size of the file in bytes
+	FileSize int64 `json:"file_size"`
+	// time of file upload
+	FileUploadTime string `json:"file_upload_time"`
 }
-func (m Integration) PrintResponse() {
+func (m SourceFileUpload) PrintResponse() {
     r, err := json.Marshal(m)
     var out bytes.Buffer
     err = json.Indent(&out, []byte(string(r)), "", "    ")