+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Rockset Go Client Changelog
 
+## v0.5.9 2019-01-23
+- Add support for AWS External ID as a new type of integration
+- Add support for Google Cloud Storage
+
 ## v0.5.8 2018-12-28
 - Support for creating collections from XML data files
 - Support for prefix patterns in S3 sources